// generated/clients/javascript/services/uigf-game-global-communityrecords/src/api/starRailChallengeApi/options.ts
// Copyright (c) Microsoft Corporation.
// Licensed under the MIT License.

import { OperationOptions } from "@azure-rest/core-client";

/** Optional parameters. */
export interface StarRailChallengeApiGetSimulatedUniverseOptionalParams extends OperationOptions {
  scheduleType?: number;
  needDetail?: boolean;
  needAll?: boolean;
}

/** Optional parameters. */
export interface StarRailChallengeApiGetApocalypticShadowOptionalParams extends OperationOptions {
  scheduleType?: number;
  needDetail?: boolean;
  needAll?: boolean;
}

/** Optional parameters. */
export interface StarRailChallengeApiGetPureFictionOptionalParams extends OperationOptions {
  scheduleType?: number;
  needDetail?: boolean;
  needAll?: boolean;
}

/** Optional parameters. */
export interface StarRailChallengeApiGetForgottenHallOptionalParams extends OperationOptions {
  scheduleType?: number;
  needDetail?: boolean;
  needAll?: boolean;
}